+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changes
 - add JSON-RPC stdio server `deltachat-rpc-server` and use it for JSON-RPC tests #3695
+- update rPGP from 0.8 to 0.9 #3737
 
 ### Fixes