+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### API-Changes
+- jsonrpc: add `miscSaveSticker` method
 
 ### Changes
 - add JSON-RPC stdio server `deltachat-rpc-server` and use it for JSON-RPC tests #3695