+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - add JSON-RPC stdio server `deltachat-rpc-server` and use it for JSON-RPC tests #3695
 - update rPGP from 0.8 to 0.9 #3737
 - jsonrpc: typescript client: use npm released deltachat fork of the tiny emitter package #3741
+- jsonrpc: show sticker image in quote #3744
 
 ### Fixes